Make test suite class names match production code names
